Initial AssertMessageToCommentConverter
xunit does not support a "message" field in its asserts.  This is
intentional: https://github.com/xunit/xunit/issues/350.

MSBuild has used the message field, though, and it seems wasteful to just
that information away.  This introduces a new converter that extracts the
message (if the extra argument in an assert is a string literal) into a
comment.