Convert ExpectedException tests to xunit
xunit does exception checking with Assert.Throws<> taking a block that is
expected to throw.  This allows very fine-grained guarantees: *that exact
operation* is the one that threw.

MSTest doesn't have this level of granularity.  It can only check that
somewhere in the body of a method, a thrown exception of a given type is
expected.

This change extends the converter to automatically include the full body
of a method with an [ExpectedException] as the body of a lambda within an
Assert.Throws call.  That replicates the MSTest behavior, so it's a nice
thing to do at automated-conversion time.

To have good xunit hygeine, a human should come along and retarget the
Assert.Throws to the guilty expression, rather than the whole method.