Make RemoveTestAttributes work more than once in a document
Before this change, the foreach loop within AddTransformation's lambda only worked for the first matching attribute in a file.  I believe this is because the remaining members of the rewrittenNodes list were not present in the (new, modified) syntax tree after the first member of the list was processed.  This wasn't caught because typically there is only one [TestClass] in a file (per the standard coding style), so this was only noticed once I started using the same code for [ExpectedException], which occurs multiple times per file.

The only difficulty in removing the attribute is that there are two possibilities for applying multiple attributes: [Attribute1,Attribute2] and [Attribute1][Attribute2]

These are now handled with different transformations, which required doing the determination of which case we're in for a given attribute early.